Update React and move permissions.js to frontend
