+++ frontend/package.json
@@ -26,7 +26,7 @@
     "react-measure": "^2.3.0",
     "react-modal": "^3.8.1",
     "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
+    "react-scripts": "3.1.2",
     "react-select": "^3.0.4",
     "react-window": "^1.8.1",
     "tippy.js": "^4.3.5",
